Minor fixes to run tests

+++ src/main/java/it/polimi/ingsw/is25am28/Components/Cannon.java
@@ -9,7 +9,7 @@ public final class Cannon extends Component {
       }
 
       public float getFirePower(){
-            return getDirection() != 0 ? force/2: force;
+            return getDirection() != 0 ? ((float) force)/2: force;
       }
 
       public boolean requireEnergy(){
